refactor: application-oauth.yml 형식에 맞게 value 값 수정
- 패키지 위치 이동
