+++ backend/src/main/java/woorifisa/goodfriends/backend/global/config/AuthenticationConfig.java
@@ -1,4 +1,4 @@
-package woorifisa.goodfriends.backend.global.config.webconfig;
+package woorifisa.goodfriends.backend.global.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -15,7 +15,7 @@ import woorifisa.goodfriends.backend.global.config.utils.JwtFilter;
 @EnableWebSecurity
 public class AuthenticationConfig {
 
-    @Value("${oauth2.jwt.secret}")
+    @Value("${security.jwt.token.secret-key}")
     private String secretKey;
 
     // SpringSecurity 적용하면 모든 api에 인증이 필요하다고 default로 호출됨
