+++ backend/src/main/java/woorifisa/goodfriends/backend/global/config/WebConfig.java
@@ -1,4 +1,4 @@
-package woorifisa.goodfriends.backend.global.config.webconfig;
+package woorifisa.goodfriends.backend.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
